\name{sabermetrics-package}
\alias{sabermetrics-package}
\alias{sabermetrics}
\docType{package}
\title{
Sabermetrics Functions For Baseball Analytics
}
\description{
A collection of baseball analytics functions for sabermetrics purposes. Among these functions include popular metrics such as wOBA, runs created functions as well as enhanced pitching metrics.
}
\details{
\tabular{ll}{
Package: \tab Sabermetrics\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2016-01-25\cr
License: \tab GPL-3\cr
}
}
\author{
Peter Xenopoulos, Fernando Crema
}
\references{
\cite{Wikipedia: http://en.wikipedia.org/wiki/Sabermetrics#Examples} |
\cite{Reddit: http://www.reddit.com/r/Sabermetrics} |
\cite{Peter Xenopoulos Website: www.peterxeno.com}
}
\keyword{ sabermetrics }
\keyword{ baseball }
\keyword{ baseball analytics }
\keyword{ sports analytics }

